refactor(basis): move public API from __init__.py into _basis_tabulate.py
Extract LagrangeVariant enum and all tabulate_* functions into the new
private module _basis_tabulate.py, then re-export them from __init__.py.
This aligns the basis package with the thin-init pattern already used
by the bspline package.
